issue-13 extract Airbrake.Payload.Backtrace from Airbrake.Payload
